move polymorphic view objects for search results to ViewComponent
