+++ app/components/search_result/base_component.html.erb
@@ -1,42 +1,39 @@
-<%# originally copied from chf_sufia app/views/curation_concerns/generic_works/_generic_work.html.erb
-    Modified for cleaner architecture using our draper-based "view models" %>
-
-<li id="document_<%= model.friendlier_id %>" class="scihist-results-list-item document <%# render_document_class generic_work %>" itemscope itemtype="<%# generic_work.itemtype %>">
+<li id="document_<%= model.friendlier_id %>" class="scihist-results-list-item document <%# render_document_class generic_work %>" itemscope itemtype="<%# generic_model.itemtype %>">
 
 
     <div class="scihist-results-list-item-thumb">
-      <%= link_to view.link_to_href do %>
-        <%= view.thumbnail_html %>
+      <%= link_to link_to_href do %>
+        <%= thumbnail_html %>
       <% end %>
-      <%= view.display_num_children %>
+      <%= display_num_children %>
     </div>
 
     <div class="scihist-results-list-item-content">
-      <% if view.display_genres.present? %>
+      <% if display_genres.present? %>
         <div class="scihist-results-list-item-genre">
-          <%= view.display_genres %>
+          <%= display_genres %>
         </div>
       <% end %>
 
       <h2 class="scihist-results-list-item-head brand-alt-h2" itemprop="name">
-        <%= link_to model.title, view.link_to_href, itemprop: "url" %>
+        <%= link_to model.title, link_to_href, itemprop: "url" %>
         <%= publication_badge(model) %>
       </h2>
 
-      <%if view.additional_title.present? %>
+      <%if additional_title.present? %>
         <div class="scihist-results-list-item-additional-titles">
             <ul>
-              <% view.additional_title.each do |title| %>
+              <% additional_title.each do |title| %>
                 <li><%= title %></li>
               <% end %>
             </ul>
         </div>
       <% end %>
 
-      <% if view.part_of_elements.present? %>
+      <% if part_of_elements.present? %>
         <div class="scihist-results-list-item-parents">
           <ul>
-            <% view.part_of_elements.each do |element| %>
+            <% part_of_elements.each do |element| %>
               <li>
                 Part of <%= element %>
               </li>
@@ -45,10 +42,10 @@
         </div>
       <% end %>
 
-      <% if view.display_dates.present? %>
+      <% if display_dates.present? %>
         <div class="scihist-results-list-item-date">
           <ul>
-            <% view.display_dates.each do |display_interval| %>
+            <% display_dates.each do |display_interval| %>
               <li>
                 <span itemprop="date_created">
                   <%= display_interval %>
@@ -61,10 +58,10 @@
 
 
 
-      <% if view.search_highlights.present? %>
+      <% if search_highlights.present? %>
         <div class="scihist-results-list-item-highlights">
           <p>
-            <%= view.search_highlights %>
+            <%= search_highlights %>
           </p>
         </div>
       <% elsif model.description.present? %>
@@ -74,9 +71,9 @@
       <% end %>
 
 
-      <% if view.metadata_labels_and_values.present? %>
+      <% if metadata_labels_and_values.present? %>
         <ul class="list-unstyled scihist-results-list-values">
-          <% view.metadata_labels_and_values.each do |field_label, field_values| -%>
+          <% metadata_labels_and_values.each do |field_label, field_values| -%>
               <li>
                 <span class="attribute-label"><%= field_label %></span>
                 <%= safe_join(field_values, ", ") %>
@@ -85,8 +82,8 @@
         </ul>
       <% end %>
 
-      <% if view.show_cart_control? %>
-        <p><%= CartControl.new(model.friendlier_id, cart_presence: view.cart_presence).display %></p>
+      <% if show_cart_control? %>
+        <p><%= CartControl.new(model.friendlier_id, cart_presence: cart_presence).display %></p>
       <% end %>
 
     </div>
