+++ spec/components/search_result/collection_component_spec.rb
@@ -1,8 +1,8 @@
 require 'rails_helper'
 
-describe CollectionResultDisplay do
+describe SearchResult::CollectionComponent, type: :component do
   let(:collection) { FactoryBot.create(:collection, contains: [create(:public_work)]) }
-  let(:rendered) { Nokogiri::HTML.fragment(described_class.new(collection, child_counter: child_counter).display) }
+  let(:rendered) { render_inline(described_class.new(collection, child_counter: child_counter)) }
 
   let(:child_counter) { ChildCountDisplayFetcher.new([collection.friendlier_id]) }
